+++ Robust.Client/UserInterface/Controls/BaseButton.cs
@@ -53,8 +53,17 @@ namespace Robust.Client.UserInterface.Controls
                 value.InternalButtons.Add(this);
                 ToggleMode = true;
 
-                // Set us to pressed if we're the first button. Doesn't go through the setter to avoid setting off our own error check.
-                _pressed = value.InternalButtons.Count == 1;
+                if (value.IsNoneSetAllowed)
+                {
+                    // Still UNPRESS if there's another pressed button, but don't PRESS it otherwise.
+                    if (value.Pressed != null)
+                        _pressed = false;
+                }
+                else
+                {
+                    // Set us to pressed if we're the first button. Doesn't go through the setter to avoid setting off our own error check.
+                    _pressed = value.InternalButtons.Count == 1;
+                }
                 DrawModeChanged();
             }
         }
@@ -99,7 +108,7 @@ namespace Robust.Client.UserInterface.Controls
                     return;
                 }
 
-                if (!value && Group != null)
+                if (!value && Group is { IsNoneSetAllowed: false })
                 {
                     throw new InvalidOperationException("Cannot directly unset a grouped button. Set another button in the group instead.");
                 }
@@ -444,6 +453,26 @@ namespace Robust.Client.UserInterface.Controls
     /// </remarks>
     public sealed class ButtonGroup
     {
+        /// <summary>
+        /// Whether it is legal for this button group to have no selected button.
+        /// </summary>
+        /// <remarks>
+        /// If true, it's legal for no button in the group to be active.
+        /// This is then the initial state of a new group of buttons (no button is automatically selected),
+        /// and it becomes legal to manually clear the active button through code.
+        /// The user cannot manually unselect the active button regardless, only by selecting a difference button.
+        /// </remarks>
+        public bool IsNoneSetAllowed { get; }
+
+        /// <summary>
+        /// Create a new <see cref="ButtonGroup"/>
+        /// </summary>
+        /// <param name="isNoneSetAllowed">The value of <see cref="IsNoneSetAllowed"/> on the new button group.</param>
+        public ButtonGroup(bool isNoneSetAllowed = false)
+        {
+            IsNoneSetAllowed = isNoneSetAllowed;
+        }
+
         internal readonly List<BaseButton> InternalButtons = new();
         public IReadOnlyList<BaseButton> Buttons => InternalButtons;