+++ RELEASE-NOTES.md
@@ -40,6 +40,8 @@ END TEMPLATE-->
 ### New features
 
 * Add an overload to DrawingHandleBase.DrawPrimitives that allows a `List<Vector2>` to be passed in.
+* `NetUserId` implements `ISelfSerialize` so can be used in data fields.
+* `ButtonGroup.IsNoneSetAllowed` to allow a button group to have no buttons pressed by default.
 
 ### Bugfixes